Add DeleteQuery class.
Split code specific to DELETE query generation into a separate class.
